fix(executor): unify is_filesystem_target to prevent safe-mode bypass via CIDR (closes #741)
